Move deprecated ImagePainter functions to a separate file.
